Bump Kotlin version. Fix deprecated stdlib dependency
The Kotlin version has been bumped from 1.2.20 to 1.2.41

The dependency on kotlin-stdlib-jre-8 has been changed to
kotlin-stdlib-jdk-8. The previous dependency was deprecated and was
changed in all other pom.xml files except for in the
vertx-lang-kotlin-coroutines directory.

The hardcoded dependency on 1.2.20 in vertx-lang-kotlin has been
changed to use the `kotlin.version` variable set in the root pom.xml
This is consistent with the pom.xml files in the other subdirectories